chore(electrostatics): export multipole PME, fix doc example refs, drop int32->long casts
- Export multipole_particle_mesh_ewald from the torch electrostatics package
  (it was defined in pme_multipole but never re-exported), matching the
  particle_mesh_ewald monopole entry.
- Fix stale gallery-example cross-references in the electrostatics user guide:
  the multipole examples are 07-10 (after the slab examples), not 05-08.
- Drop batch_idx.long() casts before bincount / scatter_add(_): torch 2.11
  accepts int32 index tensors directly, so the int32->int64 copy was dead work
  on the batched paths (multipole_autograd_batch, multipole_ewald,
  multipole_scf_step, pme_multipole).

Validated: 160 batched PME/Ewald/SCF-step tests pass; package imports clean.

+++ nvalchemiops/torch/interactions/electrostatics/multipole_autograd_batch.py
@@ -122,7 +122,7 @@ def _atom_bounds_from_batch_idx(
     tensors on the same device.
     """
     device = batch_idx.device
-    counts = torch.bincount(batch_idx.long(), minlength=batch_size)
+    counts = torch.bincount(batch_idx, minlength=batch_size)
     atom_end = counts.cumsum(dim=0).to(torch.int32)
     atom_start = torch.cat(
         [torch.zeros(1, dtype=torch.int32, device=device), atom_end[:-1]]
@@ -301,7 +301,7 @@ def _batch_rho_moment_grad_backward(
     )
     per_atom_dot = (gg_moments * grad_flat).sum(dim=-1)
     gg_volume = torch.zeros_like(volume)
-    gg_volume.scatter_add_(0, batch_idx.long(), per_atom_dot)
+    gg_volume.scatter_add_(0, batch_idx, per_atom_dot)
     gg_volume = -gg_volume / volume.detach()
     return ggrad_rho, gg_volume
